Move SelectedShapeManager into a new module
